pyproject.toml
[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[project]
name = "mu_pact_experiment"
version = "0.1.0"
description = "Reproducible µ-PACT experiments with Hydra & WandB"
requires-python = ">=3.9"
dependencies = [
    "torch>=2.1.0",
    "transformers>=4.37.0",
    "datasets>=2.14.0",
    "peft>=0.7.0",
    "accelerate>=0.24.0",
    "hydra-core>=1.3.2",
    "wandb>=0.16.0",
    "optuna>=3.5.0",
    "scipy>=1.11.0",
    "pandas>=2.1.0",
    "matplotlib>=3.8.0",
    "seaborn>=0.13.0",
    "scikit-learn>=1.3.0",
]
